sensors-logging: Refactor system so no unnecessary memory is used
The old pipeline was saving the entire log in memory, which was severely affecting the application's performance.

With this new approach, there's no more logs in memory. All log points are saved in the local database, and retrieved when the request to generate the log file is made.

I took the opportunity of this refactor to also address the problem of multiple log requesters interfering with each other. Now the system will not stop logging while there's an active log requester.